Changed quote_id's from int to long

+++ Intelipost/Intelipost.API/Model/ChangeDeliveryMethod/ChangeDeliveryMethod.cs
@@ -21,7 +21,7 @@ namespace Intelipost.API.Model
         /// ID da cotação.
         /// </summary>
         [JsonProperty("quote_id")]
-        public int? QuoteId { get; set; }
+        public long? QuoteId { get; set; }
 
         /// <summary>
         /// ID do método de envio escolhido.
